+++ src/parsing/index.ts
@@ -8,9 +8,9 @@ export function parseString(value: string): string | null {
   return value.replace(/[^A-Za-z0-9-\.]+/, "").trim() || null;
 }
 
-export function parseTags(input: string[]): string[] {
+export function parseTags(input: string): string[] {
   const tags: string[] = [];
-  input.forEach((tagString) => {
+  input.split(" ").forEach((tagString) => {
     tagString.split(",").forEach((tag) => {
       const parsedTag = parseString(tag);
       if (parsedTag && !tags.includes(parsedTag)) tags.push(parsedTag);