+++ src/parsing/session-data/index.ts
@@ -2,12 +2,13 @@ import { Result, Ok, Err } from "ts-results";
 import { bold } from "../../style";
 import { SessionBlueprint } from "../../data/session";
 import { parseTags, parseDate, parseString } from "..";
+import { parseProject } from "./project";
 
 export type SessionDataInput = {
   project: string;
   start: string;
   end: string;
-  tags: string[];
+  tags: string;
 };
 
 function errorOnNull<T>(name: string, value: T | null): Result<T, string> {
@@ -17,8 +18,8 @@ function errorOnNull<T>(name: string, value: T | null): Result<T, string> {
 export function parseSessionData(
   input: SessionDataInput
 ): Result<SessionBlueprint, string> {
-  const project = parseString(input.project);
-  if (!project) return Err(`Invalid value for ${bold("project")} name.`);
+  const projectParseResult = parseProject(input.project);
+  if (projectParseResult.err) return Err(projectParseResult.val);
 
   const dateParseResult = Result.all<Result<Date, string>[]>(
     errorOnNull("start", parseDate(input.start)),
@@ -37,7 +38,7 @@ export function parseSessionData(
   const tags = parseTags(input.tags);
 
   return Ok({
-    project,
+    project: projectParseResult.val,
     start: Math.floor(start.getTime() / 1000),
     end: Math.floor(end.getTime() / 1000),
     tags: tags.length ? tags : undefined,